Do not bump BINS in bump_bin_n_roles service
As per PEP 458 "the bins role will only be updated when a bin-n
role is added, updated, or removed". Furthermore, the bins role
is signed with "one-time keys [...] generated, used, and deleted
during the offline key ceremony."

This means bins cannot and does not have to be bumped in the
scheduled online 'bump_bin_n_roles' task.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>