Removed RolesPayload data structure
The RolesPayload was used by the TUF initialization (for development
purposes) and during the Roles Delegations.

The RolesPayload is no longer necessary during the TUF development
initialization once all the configuration is available on request
settings. During the Roles Delegations, it was replaced by the
python-TUF data structure DelegatedRole, reusing it from
``tuf.repository``.

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>